# caresoft_core/caresoft_core/Connected Services/CoreWebApi/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "caresoft_core",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Consulta/add": {
      "post": {
        "tags": [
          "Consulta"
        ],
        "operationId": "add4",
        "parameters": [
          {
            "name": "ConsultaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DocumentoPaciente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DocumentoMedico",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdConsultorio",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdAutorizacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Motivo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Comentarios",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Estado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Costo",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/update": {
      "put": {
        "tags": [
          "Consulta"
        ],
        "operationId": "update4",
        "parameters": [
          {
            "name": "ConsultaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DocumentoPaciente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DocumentoMedico",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdConsultorio",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdAutorizacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Motivo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Comentarios",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Estado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Costo",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/delete/{consultaCodigo}": {
      "delete": {
        "tags": [
          "Consulta"
        ],
        "operationId": "delete4",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Consulta/addServicio/{consultaCodigo}/{servicioCodigo}": {
      "post": {
        "tags": [
          "Consulta"
        ],
        "operationId": "addServicio",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "servicioCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/deleteServicio/{consultaCodigo}/{servicioCodigo}": {
      "delete": {
        "tags": [
          "Consulta"
        ],
        "operationId": "deleteServicio",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "servicioCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/getServicios/{consultaCodigo}": {
      "get": {
        "tags": [
          "Consulta"
        ],
        "operationId": "getServicios",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/addProducto/{consultaCodigo}/{idProducto}/{cantidad}": {
      "post": {
        "tags": [
          "Consulta"
        ],
        "operationId": "addProducto",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idProducto",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cantidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/deleteProducto/{consultaCodigo}/{idProducto}/{cantidad}": {
      "delete": {
        "tags": [
          "Consulta"
        ],
        "operationId": "deleteProducto",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idProducto",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cantidad",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consulta/getProductos/{consultaCodigo}": {
      "get": {
        "tags": [
          "Consulta"
        ],
        "operationId": "getProductos",
        "parameters": [
          {
            "name": "consultaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "loteDisponible": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ServicioDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "servicioCodigo": {
            "type": "string",
            "nullable": true
          },
          "idTipoServicio": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}